requirements.txt
numpy
ftfy==6.2.0
omegaconf==2.3.0
einops==0.8.0
torchmetrics==1.4.0.post0
pytorch-lightning==2.3.3
transformers==4.42.4
kornia==0.7.3
diffusers==0.29.2
torchvision
scipy
pandas
